+++ internal/app/wwctl/power/status/main.go
@@ -6,9 +6,9 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/warewulf/warewulf/internal/pkg/batch"
+	"github.com/warewulf/warewulf/internal/pkg/bmc"
 	"github.com/warewulf/warewulf/internal/pkg/hostlist"
 	"github.com/warewulf/warewulf/internal/pkg/node"
-	"github.com/warewulf/warewulf/internal/pkg/power"
 	"github.com/warewulf/warewulf/internal/pkg/wwlog"
 )
 
@@ -40,7 +40,7 @@ func CobraRunE(vars *variables) func(cmd *cobra.Command, args []string) (err err
 
 		batchpool := batch.New(50)
 		jobcount := len(nodes)
-		results := make(chan power.IPMI, jobcount)
+		results := make(chan bmc.TemplateStruct, jobcount)
 
 		for _, node := range nodes {
 
@@ -48,7 +48,7 @@ func CobraRunE(vars *variables) func(cmd *cobra.Command, args []string) (err err
 				wwlog.Error("%s: No IPMI IP address", node.Id())
 				continue
 			}
-			ipmiCmd := power.IPMI{
+			ipmiCmd := bmc.TemplateStruct{
 				IpmiConf: *node.Ipmi,
 				ShowOnly: vars.Showcmd,
 			}